test-tsl256x: fix type of light
The sensor value function returns
an int, so use the same type for
the light variable. This also
makes the cast to uint16_t when
printing make sense.

+++ examples/platform-specific/zoul/test-tsl256x.c
@@ -69,7 +69,7 @@ light_interrupt_callback(uint8_t value)
 PROCESS_THREAD(remote_tsl256x_process, ev, data)
 {
   PROCESS_BEGIN();
-  static uint16_t light;
+  static int light;
 
   /* Print the sensor used, teh default is the TSL2561 (from Grove) */
   if(TSL256X_REF == TSL2561_SENSOR_REF) {
